Move farmer trait to separate crate
