+++ Cargo.toml
@@ -24,6 +24,7 @@ tracing-futures = "0.2"
 sdk-utils = { path = "utils" }
 sdk-dsn = { path = "dsn" }
 sdk-node = { path = "node" }
+sdk-traits = { path = "traits" }
 
 subspace-core-primitives = { git = "https://github.com/subspace/subspace", rev = "629db2fd80120f9947916e7606d064f0f1ff23cc" }
 subspace-erasure-coding = { git = "https://github.com/subspace/subspace", rev = "629db2fd80120f9947916e7606d064f0f1ff23cc" }
@@ -154,4 +155,5 @@ members = [
     "dsn",
     "substrate",
     "node",
+    "traits"
 ]
