+++ node/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-async-trait = "0.1"
 anyhow = "1"
 backoff = "0.4"
 derivative = "2.2.0"
@@ -27,6 +26,7 @@ tracing-futures = "0.2"
 sdk-utils = { path = "../utils" }
 sdk-dsn = { path = "../dsn" }
 sdk-substrate = { path = "../substrate" }
+sdk-traits = { path = "../traits" }
 
 frame-support = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
 frame-system = { version = "4.0.0-dev", git = "https://github.com/subspace/substrate", rev = "9cf78129a2638d3f370868863d16f4fe32b4ad30" }
